#200 adjust example convention
